+++ src/postgres/statements.rs
@@ -9,7 +9,7 @@ mod update;
 
 pub use delete_from::{delete_from, DeleteFrom};
 pub use insert_into::{insert_into, BareInsertInto, InsertInto, InsertIntoColumnsBuilder, Values};
-pub use select::{select, FromItem, Joinable, OrderBy, Orderable, Select};
+pub use select::{from, select, FromItem, FromSelectBuilder, Joinable, OrderBy, Orderable, Select};
 pub use update::{update, BareUpdate, Update};
 
 pub(crate) use delete_from::delete_from_with;